--- a/internal/codeintel/dependencies/service.go
+++ b/internal/codeintel/dependencies/service.go
@@ -228,33 +228,23 @@ func (s *Service) PackagesOrVersionsMatchingFilter(ctx context.Context, filter s
 	defer endObservation(1, observation.Args{})
 
 	var (
-		matcher     packagefilters.PackageMatcher
-		nameToMatch string
-	)
-	if filter.NameFilter != nil {
-		matcher, err = packagefilters.NewPackageNameGlob(filter.NameFilter.PackageGlob)
-	} else {
-		matcher, err = packagefilters.NewVersionGlob(filter.VersionFilter.PackageName, filter.VersionFilter.VersionGlob)
-		nameToMatch = filter.VersionFilter.PackageName
-	}
-	if err != nil {
-		return nil, 0, false, errors.Wrap(err, "failed to compile glob")
-	}
-
-	var (
 		totalCount   int
 		matchingPkgs = make([]shared.PackageRepoReference, 0, limit)
 	)
 
 	if filter.NameFilter != nil {
-		var (
-			lastID    int
-			nameRegex = packagefilters.GlobToRegex(filter.NameFilter.PackageGlob)
-		)
+		// we dont use a compiled glob when checking name filters as we can do a hugely more efficient regex search
+		// in postgres instead of paging through every single package to do a glob check here
+		nameRegex, err := packagefilters.GlobToRegex(filter.NameFilter.PackageGlob)
+		if err != nil {
+			return nil, 0, false, errors.Wrap(err, "failed to compile glob")
+		}
+
+		var lastID int
 		for {
 			pkgs, _, _, err := s.store.ListPackageRepoRefs(ctx, store.ListDependencyReposOpts{
 				Scheme: filter.PackageScheme,
-				// we filter down here else we have to page through a huge number of non-matching packages
+				// we filter down here else we have to page through a potentially huge number of non-matching packages
 				Name:      reposource.PackageName(nameRegex),
 				Fuzziness: store.Regex,
 				// doing this so we don't have to load everything in at once
@@ -272,23 +262,28 @@ func (s *Service) PackagesOrVersionsMatchingFilter(ctx context.Context, filter s
 
 			lastID = pkgs[len(pkgs)-1].ID
 
+			totalCount += len(pkgs)
+
 			for _, pkg := range pkgs {
-				if matcher.Matches(pkg.Name, "") {
-					totalCount++
-					if pkg.ID <= after {
-						continue
-					}
-					if len(matchingPkgs) == limit {
-						// once we've reached the limit but are hitting more, we know theres more
-						hasMore = true
-						continue
-					}
-					pkg.Versions = nil
-					matchingPkgs = append(matchingPkgs, pkg)
+				if pkg.ID <= after {
+					continue
 				}
+				if len(matchingPkgs) == limit {
+					// once we've reached the limit but are hitting more, we know theres more
+					hasMore = true
+					continue
+				}
+				pkg.Versions = nil
+				matchingPkgs = append(matchingPkgs, pkg)
 			}
 		}
 	} else {
+		matcher, err := packagefilters.NewVersionGlob(filter.VersionFilter.PackageName, filter.VersionFilter.VersionGlob)
+		if err != nil {
+			return nil, 0, false, errors.Wrap(err, "failed to compile glob")
+		}
+		nameToMatch := filter.VersionFilter.PackageName
+
 		pkgs, _, _, err := s.store.ListPackageRepoRefs(ctx, store.ListDependencyReposOpts{
 			Scheme:    filter.PackageScheme,
 			Name:      reposource.PackageName(nameToMatch),