--- a/internal/packagefilters/BUILD.bazel
+++ b/internal/packagefilters/BUILD.bazel
@@ -2,13 +2,26 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "packagefilters",
-    srcs = ["package_filters.go"],
+    srcs = [
+        "glob.go",
+        "package_filters.go",
+    ],
     importpath = "github.com/sourcegraph/sourcegraph/internal/packagefilters",
     visibility = ["//:__subpackages__"],
     deps = [
         "//internal/codeintel/dependencies/shared",
         "//internal/conf/reposource",
         "//lib/errors",
+        "@com_github_gobwas_glob//:glob",
+    ],
+)
+
+go_test(
+    name = "glob_test",
+    timeout = "short",
+    srcs = ["glob_test.go"],
+    embed = [":glob"],
+    deps = [
         "@com_github_gobwas_glob//:glob",
     ],
 )