--- a/internal/packagefilters/BUILD.bazel
+++ b/internal/packagefilters/BUILD.bazel
@@ -1,3 +1,4 @@
+load("//dev:go_defs.bzl", "go_test")
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
@@ -13,6 +14,9 @@ go_library(
         "//internal/conf/reposource",
         "//lib/errors",
         "@com_github_gobwas_glob//:glob",
+        "@com_github_gobwas_glob//syntax",
+        "@com_github_gobwas_glob//syntax/ast",
+        "@com_github_grafana_regexp//:regexp",
     ],
 )
 
@@ -23,5 +27,15 @@ go_test(
     embed = [":glob"],
     deps = [
         "@com_github_gobwas_glob//:glob",
+    ],
+)
+
+go_test(
+    name = "packagefilters_test",
+    srcs = ["glob_test.go"],
+    embed = [":packagefilters"],
+    deps = [
+        "@com_github_gobwas_glob//:glob",
+        "@com_github_grafana_regexp//:regexp",
     ],
 )