--- a/migrations/frontend/1680296731_package_repos_package_name_regex_match/down.sql
+++ b/migrations/frontend/1680296731_package_repos_package_name_regex_match/down.sql
@@ -1,0 +1,1 @@
+DROP INDEX IF EXISTS lsif_dependency_repos_name_gin;