--- a/migrations/frontend/1680296731_package_repos_package_name_regex_match/metadata.yaml
+++ b/migrations/frontend/1680296731_package_repos_package_name_regex_match/metadata.yaml
@@ -1,0 +1,2 @@
+name: package_repos_package_name_regex_match
+parents: [1679426934, 1679432506]