--- a/migrations/frontend/1680296731_package_repos_package_name_regex_match/up.sql
+++ b/migrations/frontend/1680296731_package_repos_package_name_regex_match/up.sql
@@ -1,0 +1,3 @@
+CREATE INDEX IF NOT EXISTS lsif_dependency_repos_name_gin
+ON lsif_dependency_repos
+USING gin (name gin_trgm_ops