Name filter membership checking is currently super expensive (order of minutes on dotcom) as we page through everything every time to check if it matches in order to compute the total cost, which is way too expensive.

Instead, we can leverage being able to transform globs into regex, and GIN indexes being able to service regex lookups, to do a minimal amount of work by only ever returning the packages that match.

---

GIN index benefits arent _massive_ here given the base performance, but theyre only ~13MB on dotcom, which is a decent size for the boost

<details>
<summary>Sample query plans</summary>

```sql
EXPLAIN ANALYZE SELECT * FROM lsif_dependency_repos
WHERE name ~ '^com.sourcegraph.*$' AND scheme = 'semanticdb';
```

New

```
Bitmap Heap Scan on lsif_dependency_repos  (cost=37.60..67.65 rows=1 width=42) (actual time=0.980..1.147 rows=10 loops=1)
  Recheck Cond: (name ~ '^com.sourcegraph.*$'::text)
  Rows Removed by Index Recheck: 57
  Filter: (scheme = 'semanticdb'::text)
  Heap Blocks: exact=29
  ->  Bitmap Index Scan on noah_lsif_dependency_repos_trigram  (cost=0.00..37.60 rows=27 width=0) (actual time=0.957..0.957 rows=67 loops=1)
        Index Cond: (name ~ '^com.sourcegraph.*$'::text)
Planning Time: 0.362 ms
Execution Time: 1.206 ms
```

Old

```
Bitmap Heap Scan on lsif_dependency_repos  (cost=1320.94..10557.63 rows=1 width=42) (actual time=23.756..31.868 rows=10 loops=1)
  Recheck Cond: (scheme = 'semanticdb'::text)
  Filter: (name ~ '^com.sourcegraph.*$'::text)
  Rows Removed by Filter: 9663
  Heap Blocks: exact=835
  ->  Bitmap Index Scan on lsif_dependency_repos_scheme_id  (cost=0.00..1320.94 rows=9309 width=0) (actual time=21.740..21.741 rows=9673 loops=1)
        Index Cond: (scheme = 'semanticdb'::text)
Planning Time: 0.312 ms
Execution Time: 31.895 ms
```

```sql
EXPLAIN ANALYZE SELECT * FROM lsif_dependency_repos
WHERE name ~ '^a.*$' AND scheme = 'rust-analyzer';
```

New

```
Bitmap Heap Scan on lsif_dependency_repos  (cost=114.26..13155.52 rows=5825 width=42) (actual time=12.324..43.704 rows=7804 loops=1)
  Recheck Cond: (name ~ '^a.*$'::text)
  Rows Removed by Index Recheck: 25416
  Filter: (scheme = 'rust-analyzer'::text)
  Rows Removed by Filter: 4205
  Heap Blocks: exact=3948
  ->  Bitmap Index Scan on noah_lsif_dependency_repos_trigram  (cost=0.00..112.81 rows=13721 width=0) (actual time=11.722..11.723 rows=37425 loops=1)
        Index Cond: (name ~ '^a.*$'::text)
Planning Time: 3.330 ms
Execution Time: 44.177 ms
```

Old

```
Gather  (cost=1000.00..57949.44 rows=5825 width=42) (actual time=0.414..68.462 rows=7804 loops=1)
  Workers Planned: 2
  Workers Launched: 2
  ->  Parallel Seq Scan on lsif_dependency_repos  (cost=0.00..56366.94 rows=2427 width=42) (actual time=0.124..60.274 rows=2601 loops=3)
        Filter: ((name ~ '^a.*$'::text) AND (scheme = 'rust-analyzer'::text))
        Rows Removed by Filter: 87956
Planning Time: 0.166 ms
Execution Time: 68.887 ms
```


</details>

## Test plan

added unit tests for glob-to-regex, needs new unit tests cover matchingpackage repos